+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 * `RelayQueryField#getStorageKey` will now produce the same key regardless of the
   order of a field's arguments.
 * Range behavior keys in mutation configs are now guaranteed to be sorted.
-* Added support for the `Relay.Query` template tag which returns a
-  `RelayQuery.Root` (that can be used with `Relay.Store`).
+* Added the `Relay.createQuery()` function which returns a
+  `RelayQuery.Root` (that can be used with `Relay.Store` methods).
 
 ## 0.4.0 (October 13, 2015)